+++ tests/testthat/test-path.R
@@ -1,6 +1,13 @@
 context("Paths")
 library(xml2)
 
+style_attr <- function(nodes, attr) {
+  style <- xml_attr(nodes, "style")
+  ifelse(grepl(sprintf("%s: [^;]*;", attr), style),
+         gsub(sprintf(".*%s: ([^;]*);.*", attr), "\\1", style),
+         NA_character_)
+}
+
 test_that("paths with winding fill mode", {
   x <- xmlSVG({
     plot.new()
@@ -9,12 +16,12 @@ test_that("paths with winding fill mode", {
              col = rgb(0.5, 0.5, 0.5, 0.3), border = rgb(1, 0, 0, 0.3),
              rule = "winding")
   })
-  path <- xml_find_all(x, ".//path")
-  expect_equal(xml_attr(path, "fill-rule"), "nonzero")
-  expect_equal(xml_attr(path, "fill"), rgb(0.5, 0.5, 0.5))
-  expect_equal(xml_attr(path, "fill-opacity"), "0.30")
-  expect_equal(xml_attr(path, "stroke"), rgb(1, 0, 0))
-  expect_equal(xml_attr(path, "stroke-opacity"), "0.30")
+  path <- xml_find_one(x, ".//path")
+  expect_equal(style_attr(path, "fill-rule"), "nonzero")
+  expect_equal(style_attr(path, "fill"), rgb(0.5, 0.5, 0.5))
+  expect_equal(style_attr(path, "fill-opacity"), "0.30")
+  expect_equal(style_attr(path, "stroke"), rgb(1, 0, 0))
+  expect_equal(style_attr(path, "stroke-opacity"), "0.30")
 })
 
 test_that("paths with evenodd fill mode", {
@@ -25,12 +32,12 @@ test_that("paths with evenodd fill mode", {
              col = rgb(0.5, 0.5, 0.5, 0.3), border = rgb(1, 0, 0, 0.3),
              rule = "evenodd")
   })
-  path <- xml_find_all(x, ".//path")
-  expect_equal(xml_attr(path, "fill-rule"), "evenodd")
-  expect_equal(xml_attr(path, "fill"), rgb(0.5, 0.5, 0.5))
-  expect_equal(xml_attr(path, "fill-opacity"), "0.30")
-  expect_equal(xml_attr(path, "stroke"), rgb(1, 0, 0))
-  expect_equal(xml_attr(path, "stroke-opacity"), "0.30")
+  path <- xml_find_one(x, ".//path")
+  expect_equal(style_attr(path, "fill-rule"), "evenodd")
+  expect_equal(style_attr(path, "fill"), rgb(0.5, 0.5, 0.5))
+  expect_equal(style_attr(path, "fill-opacity"), "0.30")
+  expect_equal(style_attr(path, "stroke"), rgb(1, 0, 0))
+  expect_equal(style_attr(path, "stroke-opacity"), "0.30")
 })
 
 test_that("paths with no filling color", {
@@ -41,9 +48,12 @@ test_that("paths with no filling color", {
              col = NA, border = rgb(1, 0, 0, 0.3),
              rule = "winding")
   })
-  path <- xml_find_all(x, ".//path")
-  expect_equal(xml_attr(path, "fill-rule"), "nonzero")
-  expect_equal(xml_attr(path, "fill"), "none")
-  expect_equal(xml_attr(path, "stroke"), rgb(1, 0, 0))
-  expect_equal(xml_attr(path, "stroke-opacity"), "0.30")
+  style <- xml_text(xml_find_one(x, "//style"))
+  expect_match(style, "fill: none;")
+
+  path <- xml_find_one(x, ".//path")
+  expect_equal(style_attr(path, "fill-rule"), "nonzero")
+  expect_equal(style_attr(path, "fill"), NA_character_)
+  expect_equal(style_attr(path, "stroke"), rgb(1, 0, 0))
+  expect_equal(style_attr(path, "stroke-opacity"), "0.30")
 })