+++ tests/testthat/test-text.R
@@ -1,5 +1,12 @@
-context("text")
+context("Text")
 library(xml2)
+
+style_attr <- function(nodes, attr) {
+  style <- xml_attr(nodes, "style")
+  ifelse(grepl(sprintf("%s: [^;]*;", attr), style),
+         gsub(sprintf(".*%s: ([^;]*);.*", attr), "\\1", style),
+         NA_character_)
+}
 
 test_that("par(cex) affects strwidth", {
   devSVG(tempfile())
@@ -50,7 +57,7 @@ test_that("special characters are escaped", {
   })
   # xml_text unescapes for us - this still tests that the
   # file parses, which it wouldn't otherwise
-  expect_equal(xml_attr(xml_find_one(x, ".//text"), "fill"), "#113399")
+  expect_equal(style_attr(xml_find_one(x, ".//text"), "fill"), "#113399")
 })
 
 test_that("default point size is 12", {
@@ -58,7 +65,7 @@ test_that("default point size is 12", {
     plot.new()
     text(0.5, 0.5, "a")
   })
-  expect_equal(xml_attr(xml_find_one(x, ".//text"), "font-size"), "12.00")
+  expect_equal(style_attr(xml_find_one(x, ".//text"), "font-size"), "12.00")
 })
 
 test_that("cex generates fractional font sizes", {
@@ -66,7 +73,7 @@ test_that("cex generates fractional font sizes", {
     plot.new()
     text(0.5, 0.5, "a", cex = 0.1)
   })
-  expect_equal(xml_attr(xml_find_one(x, ".//text"), "font-size"), "1.20")
+  expect_equal(style_attr(xml_find_one(x, ".//text"), "font-size"), "1.20")
 })
 
 test_that("font sets weight/style", {
@@ -75,8 +82,8 @@ test_that("font sets weight/style", {
     text(0.5, seq(0.9, 0.1, length = 4), "a", font = 1:4)
   })
   text <- xml_find_all(x, ".//text")
-  expect_equal(xml_attr(text, "font-weight"), c(NA, "bold", NA, "bold"))
-  expect_equal(xml_attr(text, "font-style"), c(NA, NA, "italic", "italic"))
+  expect_equal(style_attr(text, "font-weight"), c(NA, "bold", NA, "bold"))
+  expect_equal(style_attr(text, "font-style"), c(NA, NA, "italic", "italic"))
 })
 
 test_that("font sets weight/style", {
@@ -87,7 +94,7 @@ test_that("font sets weight/style", {
     text(0.5, 0.9, "a", family = "mono")
   })
   text <- xml_find_all(x, ".//text")
-  expect_equal(xml_attr(text, "font-family"), c("Times New Roman", "Arial", "courier"))
+  expect_equal(style_attr(text, "font-family"), c("Times New Roman", "Arial", "courier"))
 })
 
 test_that("a symbol has width greater than 0", {
@@ -104,6 +111,6 @@ test_that("symbol font family is 'symbol'", {
     text(1, 1, expression(symbol("\042")))
   })
   text <- xml_find_all(x, ".//text")
-  expect_equal(xml_attr(text, "font-family"), c("symbol"))
+  expect_equal(style_attr(text, "font-family"), c("symbol"))
 })